2025-11-25 Version 4

+++ database/migrations/2025_11_21_104550_update_artikeleinrichtung_add_lagerort_id_and_remove_lagerort.php
@@ -8,7 +8,11 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('artikeleinrichtung', function (Blueprint $table) {
-            $table->dropColumn('lagerort');
+
+            if (Schema::hasColumn('artikeleinrichtung', 'lagerort')) {
+                    $table->dropColumn('lagerort');
+            }
+
             $table->bigInteger('lagerort_id')->nullable()->after('abladestelle_id'); // nach Bedarf positionieren
             $table->foreign('lagerort_id')->references('id')->on('lagerorte')->nullOnDelete();
         });
